requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
anthropic==0.26.0
email-validator==2.1.1
